Restructure into src folder
