+++ src/server/api/inbox.ts
@@ -3,7 +3,7 @@ import { Type } from '@sinclair/typebox'
 
 import type { APIConfig, FastifyTypebox } from '.'
 import Store from '../store'
-import type ActivityPubSystem from './apsystem.js'
+import type ActivityPubSystem from '../apsystem.js'
 
 export const inboxRoutes = (cfg: APIConfig, store: Store, apsystem: ActivityPubSystem) => async (server: FastifyTypebox): Promise<void> => {
   // Returns an JSON-LD OrderedCollection with items in the moderation queue
